Add support for building with the Clang compiler.
Set CC=clang and CXX=clang++ to use the Clang compiler.  Successful
build verified with DEBUG=1 and STRICT=1 using Clang 3.5.0.

Updated code to remove unreachable code, unused parameters, and
shadowed variables.

+++ src/hcfg.c
@@ -258,7 +258,7 @@ int hcfg_loadfile(hcfg_t* cfg, const char* filename)
         if (ptr != buf)
             memmove(buf, ptr, len);
 
-        if (len + 1 == buf_cap) {
+        if (len + 1 == (size_t)buf_cap) {
             if (array_grow(&buf, &buf_cap, sizeof(char)) != 0) {
                 perror("Error: Could not grow config parsing buffer");
                 goto error;
